Move asset_record.rs content into issuance.rs
